pyproject.toml
[tool.poetry]
name = "async-telegram-bot-gino"
version = "0.1.0"
description = ""
authors = ["VasilevAlexandr97 <vasilev.alex.work@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
gino = "^1.0.1"
aiogram = "^2.13"
uvloop = "^0.15.2"
ujson = "^4.0.2"
cchardet = "^2.1.7"
aiodns = "^3.0.0"
aiohttp = {extras = ["speedups"], version = "^3.7.4"}
python-dotenv = "^0.17.1"

[tool.poetry.dev-dependencies]
black = {version = "^21.5b1", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
